+++ .gitignore
@@ -33,7 +33,6 @@ __pycache__/
 /mingw-build/
 # Created by default with `src/ci/docker/run.sh`:
 /obj/
-/rustllvm/
 /unicode-downloads
 /target
 # Generated by compiletest for incremental:
